move examples into storybook; make example panel buttons separated by row
